Allow adding a path prefix to C# codegen options
A path prefix can now be provided in options to C# generators that will
add a prefix to all operations generated from the corresponding OpenAPI
document. This is helpful for the TestApp, but any API may be mounted at
a subpath on the server; this will help both servers and clients alike.